Move use cases from devkit to specific micro services
