+++ horusec-account/internal/handlers/companies/companies.go
@@ -25,11 +25,11 @@ import (
 	authEnums "github.com/ZupIT/horusec/development-kit/pkg/enums/auth"
 	errorsEnum "github.com/ZupIT/horusec/development-kit/pkg/enums/errors"
 	brokerLib "github.com/ZupIT/horusec/development-kit/pkg/services/broker"
-	companyUseCases "github.com/ZupIT/horusec/development-kit/pkg/usecases/company"
-	"github.com/ZupIT/horusec/development-kit/pkg/usecases/repositories"
 	httpUtil "github.com/ZupIT/horusec/development-kit/pkg/utils/http"
 	"github.com/ZupIT/horusec/horusec-account/config/app"
 	companiesController "github.com/ZupIT/horusec/horusec-account/internal/controller/companies"
+	companyUseCases "github.com/ZupIT/horusec/horusec-account/internal/usecases/company"
+	"github.com/ZupIT/horusec/horusec-account/internal/usecases/repositories"
 	"github.com/go-chi/chi"
 	"github.com/google/uuid"
 	netHttp "net/http"
