+++ horusec-account/internal/handlers/webhook/webhook_test.go
@@ -21,8 +21,8 @@ import (
 	"github.com/ZupIT/horusec/development-kit/pkg/databases/relational"
 	"github.com/ZupIT/horusec/development-kit/pkg/entities/webhook"
 	errorsEnum "github.com/ZupIT/horusec/development-kit/pkg/enums/errors"
-	webhookUseCases "github.com/ZupIT/horusec/development-kit/pkg/usecases/webhook"
 	webhookController "github.com/ZupIT/horusec/horusec-account/internal/controller/webhook"
+	webhookUseCases "github.com/ZupIT/horusec/horusec-account/internal/usecases/webhook"
 	"github.com/go-chi/chi"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
