+++ horusec-analytic/internal/usecases/dashboard/graphql_objects.go
