+++ horusec-api/internal/handlers/management/management.go
@@ -20,9 +20,9 @@ import (
 	"github.com/ZupIT/horusec/development-kit/pkg/enums/errors"
 	horusecEnums "github.com/ZupIT/horusec/development-kit/pkg/enums/horusec"
 	"github.com/ZupIT/horusec/development-kit/pkg/enums/severity"
-	managementUseCases "github.com/ZupIT/horusec/development-kit/pkg/usecases/management"
 	httpUtil "github.com/ZupIT/horusec/development-kit/pkg/utils/http"
 	"github.com/ZupIT/horusec/horusec-api/internal/controllers/management"
+	managementUseCases "github.com/ZupIT/horusec/horusec-api/internal/usecases/management"
 	"github.com/go-chi/chi"
 	"github.com/google/uuid"
 	netHTTP "net/http"
