+++ horusec-api/internal/handlers/management/management_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/ZupIT/horusec/development-kit/pkg/entities/horusec"
 	errorsEnum "github.com/ZupIT/horusec/development-kit/pkg/enums/errors"
 	horusecEnum "github.com/ZupIT/horusec/development-kit/pkg/enums/horusec"
-	managementUseCases "github.com/ZupIT/horusec/development-kit/pkg/usecases/management"
 	"github.com/ZupIT/horusec/horusec-api/internal/controllers/management"
+	managementUseCases "github.com/ZupIT/horusec/horusec-api/internal/usecases/management"
 	"github.com/go-chi/chi"
 	"github.com/stretchr/testify/assert"
 	"net/http"
